healthcare_backend/requirements.txt
django
djangorestframework
celery
python-telegram-bot
djangorestframework-simplejwt
python-decouple
psycopg2-binary
redis
python-dotenv
django-cors-headers
pillow
